move `navigation` back, `tui` to `ui` and app
