+++ src/app.rs
@@ -1,28 +1,70 @@
+use console::Key;
+use ratatui::{prelude::CrosstermBackend, Terminal};
+
+use nu_protocol::{
+    ast::{CellPath, PathMember},
+    ShellError, Span, Value,
+};
+
+use super::navigation::Direction;
+use super::{config::Config, navigation, tui};
+
 use std::error;
 
 /// Application result type.
 pub type AppResult<T> = std::result::Result<T, Box<dyn error::Error>>;
 
+/// the mode in which the application is
+#[derive(Clone, Debug, PartialEq)]
+pub(super) enum Mode {
+    /// the NORMAL mode is the *navigation* mode, where the user can move around in the data
+    Normal,
+    /// the INSERT mode lets the user edit cells of the structured data
+    Insert,
+    /// the PEEKING mode lets the user *peek* data out of the application, to be reused later
+    Peeking,
+    /// TODO: documentation
+    Bottom,
+}
+
+impl Default for Mode {
+    fn default() -> Self {
+        Self::Normal
+    }
+}
+
+impl std::fmt::Display for Mode {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        let repr = match self {
+            Self::Normal => "NORMAL",
+            Self::Insert => "INSERT",
+            Self::Peeking => "PEEKING",
+            Self::Bottom => "BOTTOM",
+        };
+        write!(f, "{}", repr)
+    }
+}
+
 /// Application.
 #[derive(Debug)]
 pub struct App {
-    /// Is the application running?
-    pub running: bool,
-    /// counter
-    pub counter: u8,
+    /// the full current path in the data
+    pub cell_path: CellPath,
+    /// the current [`Mode`]
+    pub mode: Mode,
 }
 
 impl Default for App {
     fn default() -> Self {
         Self {
-            running: true,
-            counter: 0,
+            cell_path: CellPath { members: vec![] },
+            mode: Mode::default(),
         }
     }
 }
 
 impl App {
-    /// Constructs a new instance of [`App`].
+    /// Constructs a new instance of [`State`].
     pub fn new() -> Self {
         Self::default()
     }
@@ -30,20 +72,32 @@ impl App {
     /// Handles the tick event of the terminal.
     pub fn tick(&self) {}
 
-    /// Set running to false to quit the application.
-    pub fn quit(&mut self) {
-        self.running = false;
+    pub(super) fn from_value(value: &Value) -> Self {
+        let mut state = Self::default();
+        match value {
+            Value::List { vals, .. } => state.cell_path.members.push(PathMember::Int {
+                val: 0,
+                span: Span::unknown(),
+                optional: vals.is_empty(),
+            }),
+            Value::Record { cols, .. } => state.cell_path.members.push(PathMember::String {
+                val: cols.get(0).unwrap_or(&"".to_string()).into(),
+                span: Span::unknown(),
+                optional: cols.is_empty(),
+            }),
+            _ => {}
+        }
+
+        state
     }
 
-    pub fn increment_counter(&mut self) {
-        if let Some(res) = self.counter.checked_add(1) {
-            self.counter = res;
-        }
+    /// TODO: documentation
+    pub(super) fn is_at_bottom(&self) -> bool {
+        matches!(self.mode, Mode::Bottom)
     }
 
-    pub fn decrement_counter(&mut self) {
-        if let Some(res) = self.counter.checked_sub(1) {
-            self.counter = res;
-        }
+    /// TODO: documentation
+    pub(super) fn hit_bottom(&mut self) {
+        self.mode = Mode::Bottom;
     }
 }
