+++ src/lib.rs
@@ -14,17 +14,17 @@ pub mod tui;
 pub mod handler;
 
 mod config;
+mod navigation;
 
-use app::{App, AppResult};
-use app::{Mode, State};
+use app::{App, AppResult, Mode};
 use config::Config;
 use event::{Event, EventHandler};
 use handler::handle_key_events;
 use tui::Tui;
 
-use std::io;
 use ratatui::backend::CrosstermBackend;
 use ratatui::Terminal;
+use std::io;
 
 use nu_plugin::{EvaluatedCall, LabeledError, Plugin};
 use nu_protocol::{Category, PluginExample, PluginSignature, Type, Value};
@@ -82,6 +82,16 @@ impl Plugin for Explore {
 /// 1. sets the terminal up (see [`terminal::setup`])
 /// 1. runs the application (see [`app::run`])
 /// 1. restores the terminal (see [`terminal::restore`])
+///
+/// run the application
+///
+/// this function
+/// 1. creates the initial [`State`]
+/// 1. runs the main application loop
+///
+/// the application loop
+/// 1. renders the TUI with [`tui`]
+/// 1. reads the user's input keys and transition the [`State`] accordingly
 fn explore(call: &EvaluatedCall, input: &Value) -> AppResult<Value> {
     let config = Config::from_value(call.opt(0).unwrap().unwrap_or(Value::record(
         vec![],
@@ -112,16 +122,24 @@ fn explore(call: &EvaluatedCall, input: &Value) -> AppResult<Value> {
         }
     }
 
+    // match transition_state(&key, config, &mut state, input)? {
+    //     TransitionResult { exit: true, result } => match result {
+    //         None => break,
+    //         Some(value) => return Ok(value),
+    //     },
+    //     TransitionResult { exit: false, .. } => {}
+    // }
+    // Ok(Value::nothing(Span::unknown()));
+
     // Exit the user interface.
     tui.exit()?;
 
-    match result {
-        Ok(res) => Ok(res),
-        Err(err) => Err(LabeledError {
-            label: "unexpected error".into(),
-            msg: err.to_string(),
-            span: Some(call.head),
-        }),
-    }
+    // match result {
+    //     Ok(res) => Ok(res),
+    //     Err(err) => Err(LabeledError {
+    //         label: "unexpected error".into(),
+    //         msg: err.to_string(),
+    //         span: Some(call.head),
+    //     }),
+    // }
 }
-
